fastapi/requirements.txt

pandas
chromadb
fastapi
uvicorn
tiktoken
google-generativeai